[tester] modif config

+++ srcs/Request.cpp
@@ -41,8 +41,11 @@ Request::~Request(void)
 		_fs.close();
 		if (unlink(_nameBodyFile.c_str()) == -1)
 		{
-			std::cerr << "unlink error" << std::endl;
-			std::cerr << "errno: " << strerror(errno) << std::endl;
+			if (DEBUG_REQUEST)
+			{
+				std::cerr << "unlink error" << std::endl;
+				std::cerr << "errno: " << strerror(errno) << std::endl;
+			}
 		}
 	}
 }
